move `TypeCheckResults` into its own module
