+++ components/lark-type-check/src/base_inference/query_definition.rs
@@ -1,7 +1,7 @@
 use crate::base_inference::resolve_to_base_inferred::ResolveToBaseInferred;
 use crate::base_inference::{BaseInference, BaseInferenceTables};
+use crate::results::TypeCheckResults;
 use crate::TypeCheckDatabase;
-use crate::TypeCheckResults;
 use crate::TypeChecker;
 use crate::UniverseBinder;
 use generational_arena::Arena;
