+++ components/lark-type-check/src/full_inference/type_checker.rs
@@ -6,10 +6,10 @@ use crate::full_inference::perm::PermData;
 use crate::full_inference::perm::PermVar;
 use crate::full_inference::FullInference;
 use crate::full_inference::FullInferenceTables;
+use crate::results::TypeCheckResults;
 use crate::substitute::Substitution;
 use crate::substitute::SubstitutionDelegate;
 use crate::TypeCheckDatabase;
-use crate::TypeCheckResults;
 use crate::TypeChecker;
 use crate::TypeCheckerFamilyDependentExt;
 use crate::TypeCheckerVariableExt;
